chore(dev): introduce common hook for saving new datasets
